feat(www): add Deepthi to Select VIP experience and remove dinner pages
The unused VIP dinner RSVP is gone. Deepthi is now a host on the VIP experience page with a local headshot instead of her GitHub avatar.
